Fix +1 difficulty issue, change difficulty to float, added more raid sizes
HandleStatModifier() treats 100% as a 100% buff, so SoloCraft was actually applying difficulty + 1.  Setting a difficulty to 2 was actually applying a 300% bonus.
Difficulty changed to float, to allow for fine tuning, ie. difficulty of 1.5 is now possible, rather than having no option between 1 and 2.
Raid sizes changed to support 10, 25, and 30 man raids.  Also added a modifier for mythic 5 mans.